Move datastore to common folder
