+++ tests/end_to_end/test_publish_and_read.py
@@ -4,7 +4,7 @@ from getpass import getuser
 from unittest import IsolatedAsyncioTestCase
 from paramiko import SSHClient, AutoAddPolicy, RSAKey
 from switcheroo.server.server import Server
-from switcheroo.server.data_stores import S3DataStore
+from switcheroo.data_stores import S3DataStore
 from switcheroo.publisher.key_publisher import S3Publisher
 from switcheroo import paths
